Split out core.models.instance to separate InstanceAction and InstanceStatus(+History)
